add useKeyframes, some refactor, better example app
