+++ example/public/index.html
@@ -9,9 +9,9 @@
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <title>css-system example</title>
-    __STYLE__
+    <style id="__ROOT__"></style>
   </head>
   <body>
-    <div id="root">__ROOT__</div>
+    <div id="root"><div id="__ROOT__"></div></div>
   </body>
 </html>
