+++ example/src/Input.js
@@ -1,0 +1,28 @@
+import {useCss} from "@css-system/use-css"
+import React from "react"
+
+export const Input = ({as: Component = "input", css = {}, deps, ...props}) => {
+  const className = useCss(
+    {
+      py: {_: 2, m: 3},
+      px: {_: 3, m: 4},
+      bg: "lightPrimary",
+      color: "lightPrimaryText",
+      border: "none",
+      minWidth: 0,
+      minHeight: 0,
+      flex: "none",
+      fontSize: "inherit",
+
+      ...css,
+      "&:disabled": {
+        cursor: "not-allowed",
+        opacity: 0.5,
+        ...css["&:disabled"],
+      },
+    },
+    deps
+  )
+
+  return <Component className={className} {...props} />
+}
