+++ example/src/index.js
@@ -2,4 +2,4 @@ import React from "react"
 import ReactDOM from "react-dom"
 import App from "./App"
 
-ReactDOM.hydrate(<App />, document.getElementById("root"))
+ReactDOM.render(<App />, document.getElementById("root"))
