+++ src/computeCssObject.ts
@@ -1,146 +1,6 @@
-import {SystemStyleObject, Theme, CSSObject} from "./types"
-
-const get = (obj: Object, path: string | number, def?: any) => {
-  let result = obj
-  const keys = typeof path === "string" ? path.split(".") : [path]
-  for (const key of keys) {
-    result = result ? result[key] : undefined
-  }
-  return result === undefined ? def : result
-}
-
-const aliases = {
-  bg: "backgroundColor",
-  m: "margin",
-  mt: "marginTop",
-  mr: "marginRight",
-  mb: "marginBottom",
-  ml: "marginLeft",
-  mx: "marginX",
-  my: "marginY",
-  p: "padding",
-  pt: "paddingTop",
-  pr: "paddingRight",
-  pb: "paddingBottom",
-  pl: "paddingLeft",
-  px: "paddingX",
-  py: "paddingY",
-}
-
-const multiples = {
-  marginX: ["marginLeft", "marginRight"],
-  marginY: ["marginTop", "marginBottom"],
-  paddingX: ["paddingLeft", "paddingRight"],
-  paddingY: ["paddingTop", "paddingBottom"],
-  size: ["width", "height"],
-}
-
-const scales = {
-  color: "colors",
-  backgroundColor: "colors",
-  borderColor: "colors",
-  margin: "space",
-  marginTop: "space",
-  marginRight: "space",
-  marginBottom: "space",
-  marginLeft: "space",
-  marginX: "space",
-  marginY: "space",
-  padding: "space",
-  paddingTop: "space",
-  paddingRight: "space",
-  paddingBottom: "space",
-  paddingLeft: "space",
-  paddingX: "space",
-  paddingY: "space",
-  top: "space",
-  right: "space",
-  bottom: "space",
-  left: "space",
-  gridGap: "space",
-  gridColumnGap: "space",
-  gridRowGap: "space",
-  gap: "space",
-  columnGap: "space",
-  rowGap: "space",
-  fontFamily: "fonts",
-  fontSize: "fontSizes",
-  fontWeight: "fontWeights",
-  lineHeight: "lineHeights",
-  letterSpacing: "letterSpacings",
-  border: "borders",
-  borderTop: "borders",
-  borderRight: "borders",
-  borderBottom: "borders",
-  borderLeft: "borders",
-  borderWidth: "borderWidths",
-  borderStyle: "borderStyles",
-  borderRadius: "radii",
-  borderTopRightRadius: "radii",
-  borderTopLeftRadius: "radii",
-  borderBottomRightRadius: "radii",
-  borderBottomLeftRadius: "radii",
-  borderTopWidth: "borderWidths",
-  borderTopColor: "colors",
-  borderTopStyle: "borderStyles",
-  borderBottomWidth: "borderWidths",
-  borderBottomColor: "colors",
-  borderBottomStyle: "borderStyles",
-  borderLeftWidth: "borderWidths",
-  borderLeftColor: "colors",
-  borderLeftStyle: "borderStyles",
-  borderRightWidth: "borderWidths",
-  borderRightColor: "colors",
-  borderRightStyle: "borderStyles",
-  outlineColor: "colors",
-  boxShadow: "shadows",
-  textShadow: "shadows",
-  zIndex: "zIndices",
-  width: "sizes",
-  minWidth: "sizes",
-  maxWidth: "sizes",
-  height: "sizes",
-  minHeight: "sizes",
-  maxHeight: "sizes",
-  flexBasis: "sizes",
-  size: "sizes",
-  // svg
-  fill: "colors",
-  stroke: "colors",
-}
-
-const positiveOrNegative = (scale, value) => {
-  if (typeof value !== "number" || value >= 0) {
-    return get(scale, value, value)
-  }
-  const absolute = Math.abs(value)
-  const n = get(scale, absolute, absolute)
-  if (typeof n === "string") {
-    return "-" + n
-  }
-
-  return n * -1
-}
-
-const transforms = [
-  "margin",
-  "marginTop",
-  "marginRight",
-  "marginBottom",
-  "marginLeft",
-  "marginX",
-  "marginY",
-  "top",
-  "bottom",
-  "left",
-  "right",
-].reduce(
-  (acc, curr) => ({
-    ...acc,
-    [curr]: positiveOrNegative,
-  }),
-  {}
-)
+import {aliases, multiples, scales} from "./constants"
+import {CSSObject, SystemStyleObject, Theme} from "./types"
+import {get, transforms} from "./utils"
 
 const responsive = (systemObject: SystemStyleObject, theme: Theme) => {
   const next = {}
