+++ src/index.ts
@@ -1,5 +1,6 @@
 export {useCss} from "./useCss"
 export {useGlobalCss} from "./useGlobalCss"
+export {useKeyframes} from "./useKeyframes"
 
 export {ThemeContext} from "./themeContext"
